pyproject.toml
[project]
name = "entari-plugin-database"
version = "0.1.1"
description = "Entari plugin for SQLAlchemy ORM"
authors = [
    {name = "RF-Tar-Railt", email = "rf_tar_railt@qq.com"},
]
dependencies = [
    "sqlalchemy>=2.0.42",
    "aiosqlite>=0.21.0",
    "graia-amnesia>=0.10.1",
    "tarina<0.8.0,>=0.7.1",
    "arclet-entari<0.17.0,>=0.16.1",
    "arclet-letoderea>=0.19.5",
    "alembic>=1.16.5",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "arclet-entari[full]>=0.15.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"
include = ["src/**.py"]

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["C901", "T201", "E731", "E402", "PYI055"]
